python/img-editor/pyproject.toml
[project]
name = "img-editor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.10"
dependencies = [
    "pillow>=11.3.0",
]
